Scene: Rework to add timers and light control
Now that the new elements have been introduced,
we can rework the scene.

This moves from a single button to two buttons,
each controlling a timer, both connecting to
the LightController, and some wires to show
the status.

Adjust layer orders to minimize weirdness with
overlaps.
Adjust positions and sizes of a number of elements
to reduce visual gaps.

Change folding of scene objects to hide away the
ones that the students don't really need to focus
on.