README: Update tested version and rework text
The README has not been changed in a few years and reads quite badly.
Call it an Arduino board, explain that Unity turns on the built-in LED,
highlight that this is in the context of a course, mention SerialPort and
the need for the wrapper.

Split the paragraphs a bit more aggressively and add horizontal splits
between sections.
Move the tested version up, as it is quite a useful piece of information,
and update it as the code has been ocnfirmed to work.