Move scripts not relevant for the students to "Internal"
The goal of this repository is to serve as learning tool for students,
linked to questions and exercises.
Most of the repository is "scaffolding" to this goal and not relevant
to the students and what they are learning about : communication between
a microcontroller and a game.

Move all the scripts unerlated to serial communication or the TCP example
into an "Internal" folder, to highlight what is relevant for the students.